Namespace import commands under enum Import (#44)
Wrap the three import command types as nested types of a caseless
`enum Import`, keeping the `Command` suffix:

  ImportPodcastCommand   -> Import.PodcastCommand
  ImportMailchimpCommand -> Import.MailchimpCommand
  ImportWordPressCommand -> Import.WordPressCommand

The `enum Import {}` namespace is declared once in Import.swift; each
command is added via `extension Import` in its own renamed file
(Import.{Podcast,Mailchimp,WordPress}Command.swift). CLI command names
("import podcast", "import mailchimp", "import wordpress") and all
behavior are unchanged. CommandDispatcher registration updated to the
nested type names. PublishCommand, EpisodeURLCommand, CommandError,
CommandDispatcher, and ImportSupport remain top-level.
